% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPSRMpF.R
\name{getQdata}
\alias{getQdata}
\title{return quantitative LFQ data for UPS group}
\usage{
getQdata(res_l, UPSname_s)
}
\arguments{
\item{res_l}{(loaded MaxQuant data)}

\item{UPSname_s}{(name of UPS group - see showUPSgroups)}
}
\value{

}
\description{
return quantitative LFQ data for UPS group
}
\examples{
---
}
