% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPSRMpF.R
\name{getUbQdata}
\alias{getUbQdata}
\title{Return quantitative data for a UbL}
\usage{
getUbQdata(res_l, UPSname_s, Fac = NULL, comp = NULL, filename = NULL)
}
\arguments{
\item{res_l}{(MaxQuant result)}

\item{UPSname_s}{(see showUBls)}

\item{Fac}{(factor - if set to null only quantitative data is returned)}

\item{comp}{(comparisons to make e.g. list( c("MG", "DMSO"), c("PR", "DMSO")))}

\item{filename}{\subsection{(filename for PDF output)}{
}}
}
\description{
Return quantitative data for a UbL
}
