% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPSRMpF.R
\name{showGenes}
\alias{showGenes}
\title{Return gene names for the specific UPS group}
\usage{
showGenes(name_s)
}
\arguments{
\item{name_s}{\subsection{(name of UPS group - see showUPSgroups)}{
}}
}
\description{
Return gene names for the specific UPS group
}
