% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPSRMpF.R
\name{showUBls}
\alias{showUBls}
\title{\subsection{Return possible ubiquitin and ubiquitin like modifiers - to use with getUbQdata}{
}}
\usage{
showUBls()
}
\description{
\subsection{Return possible ubiquitin and ubiquitin like modifiers - to use with getUbQdata}{
}
}
