% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPSRMpF.R
\name{showUPSgroups}
\alias{showUPSgroups}
\title{\subsection{return the names for possible UPS groups (to use with getQdata)}{
}}
\usage{
showUPSgroups()
}
\description{
\subsection{return the names for possible UPS groups (to use with getQdata)}{
}
}
